src/backend/requirements.txt
fastapi
langchain
langchain-core
chromadb
sentence-transformers
transformers
torch
chromadb
python-jose[cryptography] 
passlib==1.7.4
bcrypt==4.0.1 
sqlalchemy
pydantic-settings
pymysql
uvicorn